requirements.txt
Flask
flask-cors
torch
torchvision
Pillow
matplotlib
gunicorn
numpy
opencv-python
azure-storage-blob
flask-cors
opencv-python-headless